+++ src/sections/collection-featured-items/featured-items-form/featured-item-field/ImageField.tsx
@@ -4,7 +4,7 @@ import { Controller, UseControllerProps, useFormContext } from 'react-hook-form'
 import { Button, Col, Form, Stack, Tooltip } from '@iqss/dataverse-design-system'
 import { ArrowDownUp, ArrowCounterclockwise, XLg } from 'react-bootstrap-icons'
 import cn from 'classnames'
-import styles from './FeaturedItem.module.scss'
+import styles from './FeaturedItemField.module.scss'
 
 interface ImageFieldProps {
   itemIndex: number
